fix: emoji's not showing in storage bug
