+++ packages/client/lib/app/core/widgets/cards/content_card/content_card_store.dart
